refactor(87b): extract shared _AliasRepositoryBase to dedup alias/tag_alias repos
AliasRepository and TagAliasRepository were byte-for-byte mirrors of 12
methods modulo three axes (table name, SQL alias token, record class).
Extract them into a private Generic base _AliasRepositoryBase[T]:

- core/database/_alias_base.py: 12 shared methods, bodies verbatim with
  only self._table / self._sql_alias / self._record_cls parametrization.
  The 3 json_each f-strings are char-identical to the originals modulo
  the alias token. Base only does `from . import connection` (no cycle).
- alias.py / tag_alias.py: subclasses set 3 class attrs, drop the 12
  bodies. AliasRecord/TagAliasRecord dataclasses and sync_from_favorite
  (asymmetric) stay put. CD-58-3 own-table-only semantics preserved
  structurally via _table; rationale relocated to TagAliasRepository
  class docstring.
- facade __init__.py unchanged; _AliasRepositoryBase stays private.

Zero behavior change: full suite 4848 passed / 2 skipped (== pre-dedup
baseline), guarded by the T-1 cross-repo equivalence net. Net -549 lines.
